Move select to a separate folder
